test(integration): dedup user-story-3 temp_project fixture into conftest
Both test_user_story_3_extract_integration.py and
test_user_story_3_query_integration.py contained a complete copy of the
same ~440-line inline source fixture (ComplexService.java ~157 lines,
DataManager.ts ~127 lines, analytics_engine.py ~156 lines).

Create tests/integration/mcp/conftest.py with a shared temp_project
fixture that is the union of both originals:
  - Path(temp_dir).resolve() from the extract side (macOS symlink fix)
  - docs/API_Reference.md from the query side

Remove the duplicated fixture definitions from both test files.
The tools fixture (small, class-scoped) stays in each file.
No test behaviour is changed.
